requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tensorflow==2.13.0
opencv-python==4.8.0
pillow==10.0.0
jupyter==1.0.0 